refactor: use tsup to bundle theme
