--- packages/theme/vite.config.ts
@@ -1,27 +1,0 @@
-import { resolve } from 'node:path';
-
-import { vanillaExtractPlugin } from '@vanilla-extract/vite-plugin';
-import { fileURLToPath } from 'url';
-import { defineConfig } from 'vite';
-import dts from 'vite-plugin-dts';
-
-const __dirname = fileURLToPath(new URL('.', import.meta.url));
-
-export default defineConfig({
-  build: {
-    terserOptions: {
-      ecma: 2020,
-    },
-    sourcemap: true,
-    lib: {
-      entry: {
-        index: resolve(__dirname, 'src/index.ts'),
-        css: resolve(__dirname, 'src/index.css.ts'),
-        'v2/index': resolve(__dirname, 'src/v2/index.ts'),
-        'presets/typography': resolve(__dirname, 'src/presets/typography.css.ts'),
-      },
-      name: 'ToEverythingTheme',
-    },
-  },
-  plugins: [dts(), vanillaExtractPlugin()],
-});
